Extract setup reconciliation orchestration
